Add support for capturing viewport when rendering images
Adds a new 'onlyViewport' option to the render API so that images can be captured
based on the viewport size when set to 'true' as opposed to being
captured as a page based on the content size when set to 'false'. The
'onlyViewport' option defaults to 'false' for backward compatibility.

Fixes #10619

Align on SlimerJS